remove external conf loaders
