+++ common/cmdarg/errors.generated.go
@@ -1,4 +1,4 @@
-package jsonem
+package cmdarg
 
 import "v2ray.com/core/common/errors"
