+++ main/yaml/yaml.go
@@ -1,39 +1,69 @@
 package yaml
 
-//go:generate go run v2ray.com/core/common/errors/errorgen
-
 import (
-	"fmt"
+	"bytes"
+	"errors"
 	"io"
-	"os"
+	"io/ioutil"
 
 	"v2ray.com/core"
 	"v2ray.com/core/common"
 	"v2ray.com/core/common/cmdarg"
-	"v2ray.com/core/main/confloader"
+	"v2ray.com/core/infra/conf/json"
+	"v2ray.com/core/infra/conf/merge"
+	"v2ray.com/core/infra/conf/serial"
 )
 
 func init() {
 	common.Must(core.RegisterConfigLoader(&core.ConfigFormat{
 		Name:      []string{"YAML"},
-		Extension: []string{".yaml", ".yml"},
+		Extension: []string{".yml", ".yaml"},
 		Loader: func(input interface{}) (*core.Config, error) {
 			switch v := input.(type) {
 			case cmdarg.Arg:
-				r, err := confloader.LoadExtConfig(append([]string{"-input=yaml", "-output=protobuf"}, v...), os.Stdin)
+				bs, err := yamlsToJSON(v)
 				if err != nil {
-					return nil, fmt.Errorf("failed to execute v2ctl to convert config file: %s", err)
+					return nil, err
 				}
-				return core.LoadConfig("protobuf", "", r)
+				data, err := merge.BytesToJSON(bs)
+				if err != nil {
+					return nil, err
+				}
+				r := bytes.NewReader(data)
+				cf, err := serial.DecodeJSONConfig(r)
+				if err != nil {
+					return nil, err
+				}
+				return cf.Build()
 			case io.Reader:
-				r, err := confloader.LoadExtConfig([]string{"--input=yaml", "-output=protobuf", "stdin:"}, os.Stdin)
+				bs, err := ioutil.ReadAll(v)
 				if err != nil {
-					return nil, fmt.Errorf("failed to execute v2ctl to convert config file: %s", err)
+					return nil, err
 				}
-				return core.LoadConfig("protobuf", "", r)
+				bs, err = json.FromYAML(bs)
+				if err != nil {
+					return nil, err
+				}
+				return serial.LoadJSONConfig(bytes.NewBuffer(bs))
 			default:
-				return nil, fmt.Errorf("unknown type")
+				return nil, errors.New("unknow type")
 			}
 		},
 	}))
+}
+
+func yamlsToJSON(files []string) ([][]byte, error) {
+	jsons := make([][]byte, 0)
+	for _, file := range files {
+		bs, err := cmdarg.LoadArgToBytes(file)
+		if err != nil {
+			return nil, err
+		}
+		j, err := json.FromYAML(bs)
+		if err != nil {
+			return nil, err
+		}
+		jsons = append(jsons, j)
+	}
+	return jsons, nil
 }
